Switched to new hardware - RPi3b --> RPi4b
The RPi3b did not perform smoothly so decided to change that. It did effect
the .ino sketches. One of the 'errors' is that the Arduino IDE on the RPi will not handle sketch filenames starting with a number. Zo 2Wire... is changed to 2Wire...
The project itself did not change.