rewrite write_serial.ino
Had to step back and investigate the strange LED behaviour (LOW and HIGH reversed)
After rebuilding and adding several 'debug' lines checking status of commands, pins and statussus the LEDs seems to work as expected.
Only the ALL command status does not behave as expected.

TODO - Check what is wrong in the if - else if  - else structure.