Added Temperature web app files AB#94
In this repository only the temp_WebApp.py and the temp_WebApp Arduino folder are added.
The PHP file tempApp.php is located on another location. The /usr/www/html folder.
This is also the location of the data.txt file. This location is not part of the repository yet.

TODO: figure out how to add this part of the environment to the repository.