Added the Web Based Fan Switch project
User story AB#98 and task AB#97. Besides these controller and backend parts,
the project contains a PHP frontend. Located in /var/www/html. Until this
moment not sure how to organize the version control of this front end and
how to keep it in sync with the back end developments. Two possible strategies to consider:
 - Create one big repository
 - Create a separate repo for the front end